backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.5.3
celery==5.3.6
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
litellm==1.17.0
python-dotenv==1.0.1
httpx==0.26.0
pytest==8.0.0
bcrypt==4.0.1
redis==5.0.1
sentry-sdk[fastapi]==1.40.6
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
python-json-logger
playwright
beautifulsoup4
openai
structlog
prometheus-fastapi-instrumentator
python-socketio==5.11.0
pyotp
qrcode[pil]
pywebpush
stripe
chromadb
sentence-transformers
transformers<4.40.0
